refactor: move routes to a separate file
